stable-baselines3/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
stable-baselines3==2.3.2
gymnasium==0.29.1
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
waitress==2.1.0



